Remove std::vector usage from command args
Avoid dynamic allocations and remove strict container type for the interface.

Using std::span ultimately means that the consumer will no longer be forced to use
a vector as data container.

By using a std::array as container we avoid dynamic allocations.
A limitation is the buffer size which is now an argument when configuring Yash.
Could be deduced at compile time but it requires that the command list is known
at compile time which is yet another update to the interface. Might follow.

Also updated interface to utilize std::string_view instead as the strings were referenced anyways.